+++ packages/kue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @idearium/kue
 
+## Unreleased
+
+### Changed
+
+-   Added more error logging.
+
 ## v1.0.0-beta.2 - 2021-10-07
 
 ### Changed